refactor(ingest): move _cleanup_saved_upload to the ingest service
All three upload doors use it — direct, presigned completion and URL
import — so it belongs beside the drain it awaits and the save it undoes
rather than in the router that hosts two of them. Extracting the
URL-import staging cluster into its own module needs it reachable from
both sides without a module-level cycle, which this gives it.

The helper resolves the storage provider in its own S3 branch, the way
save_upload_file and _download_to_file_draining in the same module do, so
the suites that patched the provider on the router patch it at the
provider module instead.

Refs #1711
